requirements.txt
asgiref==3.3.1
assertpy==1.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
djangorestframework-mango==0.1.1
idna==2.10
jsonschema==3.2.0
model-bakery==1.2.1
numpy==1.19.5
psycopg2==2.8.6
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
semantic-version==2.8.5
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
